+++ Amazon.Pay.API.SDK.Tests/ApiConfigurationTests.cs
@@ -77,5 +77,51 @@ namespace Amazon.Pay.API.SDK.Tests
                 privateKey: privateKey 
             );
         }
+
+        [Test]
+        public void CanInstantiateApiConfigurationForEnvironmentSpecificPublicKeyId()
+        {
+            // Fake Public keys
+            const string livePublicKeyId = "LIVE-XXXXXXXXXXXXXXXXXXXXXXXX";
+            const string sandboxPublicKeyId = "SANDBOX-XXXXXXXXXXXXXXXXXXXXXXXX";
+
+            // Testing Live PublicKeyId for region UnitedStates
+            AssertEnvironmentSpecificPublicKeyId(Region.UnitedStates, livePublicKeyId);
+
+            // Testing Live PublicKeyId for region Europe
+            AssertEnvironmentSpecificPublicKeyId(Region.Europe, livePublicKeyId);
+
+            // Testing Live PublicKeyId for region Japan
+            AssertEnvironmentSpecificPublicKeyId(Region.Japan, livePublicKeyId);
+
+            // Testing Sandbox PublicKeyId for region UnitedStates
+            AssertEnvironmentSpecificPublicKeyId(Region.UnitedStates, sandboxPublicKeyId);
+
+            // Testing Sandbox PublicKeyId for region Europe
+            AssertEnvironmentSpecificPublicKeyId(Region.Europe, sandboxPublicKeyId);
+
+            // Testing Sandbox PublicKeyId for region Japan
+            AssertEnvironmentSpecificPublicKeyId(Region.Japan, sandboxPublicKeyId);
+        }
+
+        // Generic methods to assert environment specific publicKeyId
+        private void AssertEnvironmentSpecificPublicKeyId(Region region, string publicKeyId)
+        {
+            // Configuration
+            var payConfig = new ApiConfiguration
+            (
+                region: region,
+                publicKeyId: publicKeyId,
+                privateKey: "-----BEGIN RSA PRIVATE KEY-----" // Fake Private key
+            );
+
+            // Assertion
+            Assert.NotNull(payConfig);
+            Assert.AreEqual(region, payConfig.Region);
+            Assert.AreEqual(publicKeyId, payConfig.PublicKeyId);
+            Assert.AreEqual(3, payConfig.MaxRetries);
+            Assert.AreEqual(Constants.ApiVersion, payConfig.ApiVersion);
+            payConfig.PrivateKey.Should().StartWith("-----BEGIN RSA");
+        }
     }
 }